form

+++ src/Controller/AnnounceController.php
@@ -28,10 +28,14 @@ use Symfony\Component\Mime\Email;
 use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
+use Symfony\Component\Security\Core\Security;
 use Symfony\Component\Security\Http\Authentication\UserAuthenticatorInterface;
 use Symfony\Component\String\Slugger\SluggerInterface;
 use Symfony\Component\Workflow\Exception\LogicException;
 use Symfony\Component\Workflow\WorkflowInterface;
+use Facebook\Facebook;
 
 class AnnounceController extends AbstractController
 {
@@ -47,20 +51,20 @@ class AnnounceController extends AbstractController
     #[Route('/myannounce/{id}', name: 'app_announce')]
     public function index(int $id): Response
     {
-        // Récupérer l'utilisateur connecté
+        // Récupére l'utilisateur connecté
         $user = $this->getUser();
 
-        // Vérifier si l'utilisateur est connecté
+        // Vérifie si l'utilisateur est connecté
         if ($user === null) {
             // Rediriger vers une page d'erreur ou retourner une réponse appropriée
             // en cas d'utilisateur non connecté
             return $this->render('utilisation/error.html.twig');
         }
 
-        // Récupérer les annonces de l'utilisateur
+        // Récupère les annonces de l'utilisateur
         $announces = $user->getAnnounces();
 
-        // Passer les annonces à la vue pour les afficher
+        // Passe les annonces à la vue pour les afficher
         return $this->render('announce/index.html.twig', [
             'announces' => $announces,
         ]);
@@ -72,30 +76,27 @@ class AnnounceController extends AbstractController
 
         $user = $this->getUser();
 
-        // Récupérer l'annonce à partir de l'identifiant
+        // Récupère l'annonce à partir de l'identifiant
         $announce = $entityManager->getRepository(Announce::class)->find($id);
 
-        // Vérifier si l'annonce existe
+        // Vérifie si l'annonce existe
         if (!$announce) {
             throw $this->createNotFoundException('Annonce non trouvée');
         }
 
-        // Vérifier si l'utilisateur est propriétaire de l'annonce
+        // Vérifie si l'utilisateur est propriétaire de l'annonce
         if ($this->getUser() !== $announce->getUser()) {
             throw $this->createAccessDeniedException('Vous n\'êtes pas autorisé à supprimer cette annonce.');
         }
 
-        // Supprimer l'annonce
+        // Supprime l'annonce
         $entityManager->remove($announce);
         $entityManager->flush();
 
-        // Rediriger vers la page "Mes annonces" avec un message de succès
+        // Redirige vers la page "Mes annonces" avec un message de succès
         $this->addFlash('success', 'L\'annonce a été supprimée avec succès.');
         return $this->redirectToRoute('app_announce', ['id' => $user->getId()]);
     }
-
-
-
 
 
     #[Route('/new', name: 'app_new')]
@@ -189,12 +190,15 @@ class AnnounceController extends AbstractController
 
 
     #[Route('/announce/create', name: 'app_announce_create')]
-    public function create(Request $request, SessionInterface $session): Response
+    #[Route("/announce/update/{id}", name: 'update_announce')]
+    public function create(Request $request, SessionInterface $session, Announce $announce = null): Response
     {
         $catId = $session->get('cat_id');
         $user = $this->getUser();
 
+        if (!$announce){
         $announce = new Announce();
+        }
 
         $announce->setUser($user);
         $announce->setCat($this->entityManager->getRepository(Cat::class)->find($catId));
@@ -204,30 +208,22 @@ class AnnounceController extends AbstractController
 
         if ($form->isSubmitted() && $form->isValid()) {
             $announce = $form->getData();
-
             $user = $this->getUser();
-
-            // Associate the current user with the announcement
             $announce->setUser($user);
-
             $announce->setCountry('France');
-
             $this->entityManager->persist($user);
 
             try {
                 $this->announceWorkflow->apply($announce, 'to_online');
             } catch (\LogicException $exception) {
-
             }
 
             $this->entityManager->persist($announce);
             $this->entityManager->flush();
-
-            // Redirection vers l'annonce crée
             $announceId = $announce->getId();
-
             return $this->redirectToRoute('app_announce_show', ['id' => $announceId]);
         }
+
         return $this->render('announce/create.html.twig', [
             'form' => $form->createView(),
         ]);
@@ -278,8 +274,20 @@ class AnnounceController extends AbstractController
 
             $user = $this->getUser();
 
+            // Envoie une notification par email à l'utilisateur de l'annonce
+            $announceUser = $announce->getUser();
+            if ($announceUser) {
+                $email = (new Email())
+                    ->from('contactez.ronron@gmail.com')
+                    ->to($announceUser->getEmail())
+                    ->subject('Nouveau commentaire sur votre annonce de chat perdu')
+                    ->html('<p>Bonjour,</p><p>Un utilisateur a posté un nouveau commentaire sur votre annonce de chat perdu : "'.$announce->getTitle().'".</p><p>Consultez votre annonce pour lire le commentaire.</p><p>Merci de votre utilisation de notre site.</p>');
 
-            // Envoyer un email de signalement à l'administrateur
+                $mailer->send($email);
+            }
+
+
+            // Envoie un email de signalement à l'administrateur
             $email = (new Email())
                 ->from($user->getEmail())
                 ->to('contactez.ronron@gmail.com')
@@ -296,7 +304,7 @@ class AnnounceController extends AbstractController
 
         if ($reportAnnounce->isSubmitted() && $reportAnnounce->isValid()) {
 
-            // Envoyer la notification à l'administrateur
+            // Envoie la notification à l'administrateur
             $reason = $reportAnnounce->get('reason')->getData();
             $details = $reportAnnounce->get('details')->getData();
 
@@ -306,7 +314,7 @@ class AnnounceController extends AbstractController
                 throw $this->createNotFoundException('Annonce non trouvée');
             }
             $user = $this->getUser();
-            // Envoyer un email de signalement à l'administrateur
+            // Envoie un email de signalement à l'administrateur
             $email = (new Email())
                 ->from($user->getEmail())
                 ->to('contactez.ronron@gmail.com')
@@ -336,40 +344,87 @@ class AnnounceController extends AbstractController
     #[Route('/announce/{id}/change_state/{state}', name: 'app_announce_change_state')]
     public function changeState(Announce $announce, string $state): Response
     {
-        // Vérifier que l'utilisateur connecté est le propriétaire de l'annonce
+        // Vérifie que l'utilisateur connecté est le propriétaire de l'annonce
         $user = $this->getUser();
         if (!$announce->getUser() || $announce->getUser()->getId() !== $user->getId()) {
             throw new AccessDeniedHttpException("Vous n'êtes pas autorisé à changer l'état de cette annonce.");
         }
 
-        // Vérifier que l'état demandé est une transition valide
+        // Vérifie que l'état demandé est une transition valide
         if (!$this->announceWorkflow->can($announce, $state)) {
             throw new BadRequestHttpException("La transition $state n'est pas autorisée pour cette annonce.");
         }
 
-        // Appliquer la transition demandée sur l'annonce
+        // Applique la transition demandée sur l'annonce
         try {
             $this->announceWorkflow->apply($announce, $state);
         } catch (LogicException $e) {
             throw new BadRequestHttpException("Impossible de changer l'état de l'annonce : " . $e->getMessage());
         }
 
-        // Enregistrer les modifications dans la base de données
+        // Enregistre les modifications dans la base de données
         $this->entityManager->flush();
 
-        // Rediriger vers la page de l'annonce modifiée
+        // Redirige vers la page de l'annonce modifiée
         return $this->redirectToRoute('app_announce_show', ['id' => $announce->getId()]);
     }
 
 
-    #[Route('/delete/{id}', name: 'app_comment_delete')]
-    public function delete(Request $request, Comment $comment): Response
+    #[Route('/delete/{id}', name: 'app_comment_delete')] // Définit l'URL pour cette fonction
+    public function delete(Request $request, Comment $comment, Security $security): Response
     {
+        // Vérifie que l'utilisateur est connecté et qu'il a les autorisations nécessaires pour supprimer le commentaire
+        if (!$security->isGranted('ROLE_USER')) {
+            throw new AccessDeniedException('Vous devez être connecté pour supprimer un commentaire.');
+        }
+
+        // Vérifie si l'utilisateur connecté est le propriétaire du commentaire
+        if ($comment->getUser() !== $security->getUser()) {
+            throw new AccessDeniedException('Vous n\'êtes pas autorisé à supprimer ce commentaire.');
+        }
+
+        // Supprime le commentaire en utilisant Doctrine EntityManager
         $this->entityManager->remove($comment);
         $this->entityManager->flush();
 
+        // Ajoute un message flash pour informer l'utilisateur que le commentaire a été supprimé avec succès
         $this->addFlash('message', 'Votre commentaire a bien été supprimé');
 
+        // Redirige l'utilisateur
         return $this->redirect($request->headers->get('referer'));
     }
+
+
+
+    /**
+     * @Route("/announce/{id}/share/facebook", name="announce_share_facebook")
+     */
+    public function shareOnFacebook(EntityManagerInterface $entityManager, Request $request, $id): \Symfony\Component\HttpFoundation\RedirectResponse
+    {
+        // Récupére l'annonce à partager
+        $announce = $entityManager->getRepository(Announce::class)->find($id);
+
+        // Vérifie si l'annonce existe
+        if (!$announce) {
+            throw $this->createNotFoundException('L\'annonce n\'existe pas.');
+        }
+
+        // Configure le SDK Facebook
+        $fb = new Facebook([
+            'app_id' => $_ENV['FACEBOOK_APP_ID'],
+            'app_secret' => $_ENV['FACEBOOK_APP_SECRET'],
+            'default_graph_version' => 'v11.0',
+        ]);
+
+        // Crée l'URL de partage sur Facebook
+        $shareUrl = $request->getSchemeAndHttpHost() . $this->generateUrl('app_announce_show', ['id' => $id], UrlGeneratorInterface::ABSOLUTE_URL);
+        $shareParams = [
+            'link' => $shareUrl,
+        ];
+        $loginUrl = $fb->getRedirectLoginHelper()->getLoginUrl($shareUrl, ['publish_actions'], http_build_query($shareParams, '', '&', PHP_QUERY_RFC3986));
+
+        // Redirige vers l'URL de partage sur Facebook
+        return $this->redirect($loginUrl);
+    }
+
 }
